Refactor models.py into folders and remove unused Job model
