+++ backend/api/models/__init__.py
@@ -1,0 +1,5 @@
+"""All Django models available"""
+
+from backend.api.models.datasets import Collection, DatasetType, Dataset, DatasetFile
+from backend.api.models.metadata import AudioMetadatum, GeoMetadatum, SpectroConfig, TabularMetadatum, TabularMetadataVariable, TabularMetadataShape
+from backend.api.models.annotations import AnnotationTag, AnnotationSet, AnnotationCampaign, AnnotationResult, AnnotationSession, AnnotationTask
